pyproject.toml
[tool.poetry]
name = "fastapi-task-manager"
version = "0.1.1"
description = "FastAPI Task Manager"
authors = ["VintlGvard"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
Django = "^5.1.3"
djangorestframework = "3.15.2"
drf-yasg = "1.21.8"
inflection = "0.5.1"
modeltranslation = "0.25"
packaging = "24.2"
pytz = "2024.2"
PyYAML = "6.0.2"
tzdata = "2024.2"
uritemplate = "4.1.1"
channels = {extras = ["daphne"], version = "^4.2.0"}
psycopg2 = "2.9.10"
django-filter = "24.3"
djangorestframework-simplejwt = "^5.3.1"
python-jose = "^3.3.0"
cryptography = "^43.0.3"
pillow = "^11.0.0"
drf-spectacular = "^0.27.2"
pytest = "^8.3.3"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
